// src/app/Models/ClientesReques.ts
export interface ClientesRequest {
    nombres: string;
    apellidos: string;
    dni: string;
    edad: number;
    status: boolean;
  }


  export interface ClientesEdiarRequest {
    id: number
    nombres: string;
    apellidos: string;
    dni: string;
    edad: number;
    status: boolean;
  }